+++ src/test/java/com/picimako/justkitting/action/JustKittingActionTestBase.java
@@ -1,5 +1,4 @@
 //Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
-
 package com.picimako.justkitting.action;
 
 import com.intellij.openapi.actionSystem.AnAction;